Further improved the circuit, especially regarding the relais circuit.
Now 10 releais are used, this enables separate measurement of all possible
current/pin combinations. Therefor an analysis of every single pin is
possible.
For now only a serial connection is implemented to transfer the result
to a connected PC. As a next step the possibility to implement a LED based
stand alone display.
I started to create a basic firmware structure which is empty for now.